Backend/requirements.txt
dj-database-url==0.5.0
Django==2.2.3
django-cors-headers==3.0.2
django-filter==2.2.0
django-heroku==0.3.1
django-resized==0.3.9
django-rest-passwordreset==1.0.0
django-sendgrid-v5==0.8.0
djangorestframework==3.10.2
future==0.17.1
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
python-http-client==3.1.0
pytz==2019.1
sendgrid==6.0.5
sqlparse==0.3.0
whitenoise==4.1.3
